♻️ Move RowsOptions and DefaultRowsOptions to opts
