+++ rows.go
@@ -2,25 +2,18 @@ package sqlslog
 
 import (
 	"database/sql/driver"
-	"errors"
-	"io"
 	"log/slog"
 	"reflect"
+
+	"github.com/akm/sql-slog/opts"
 )
 
-type RowsOptions struct {
-	Close         *StepOptions
-	Next          *StepOptions
-	NextResultSet *StepOptions
-}
+type RowsOptions = opts.RowsOptions
 
-func DefaultRowsOptions(formatter StepLogMsgFormatter) *RowsOptions {
-	return &RowsOptions{
-		Close:         DefaultStepOptions(formatter, "Rows.Close", LevelDebug),
-		Next:          DefaultStepOptions(formatter, "Rows.Next", LevelDebug, HandleRowsNextError),
-		NextResultSet: DefaultStepOptions(formatter, "Rows.NextResultSet", LevelDebug),
-	}
-}
+var (
+	DefaultRowsOptions  = opts.DefaultRowsOptions
+	HandleRowsNextError = opts.HandleRowsNextError
+)
 
 func WrapRows(original driver.Rows, logger *StepLogger, options *RowsOptions) driver.Rows {
 	if original == nil {
@@ -143,18 +136,4 @@ func (r *rowsNextResultSetWrapper) NextResultSet() error {
 			WithNilAttr(r.original.NextResultSet),
 		),
 	)
-}
-
-// HandleRowsNextError returns a boolean indicating completion and a slice of slog.Attr.
-// If err is nil, it returns true and a slice of slog.Attr{slog.Bool("eof", false)}.
-// If err is io.EOF, it returns true and a slice of slog.Attr{slog.Bool("eof", true)}.
-// Otherwise, it returns false and nil.
-func HandleRowsNextError(err error) (bool, []slog.Attr) {
-	if err == nil {
-		return true, []slog.Attr{slog.Bool("eof", false)}
-	}
-	if errors.Is(err, io.EOF) {
-		return true, []slog.Attr{slog.Bool("eof", true)}
-	}
-	return false, nil
 }
